FIXED - Bug from user navagating prior to signin
